Match custom Cloud VM menu row metrics
